Clear client expedition info immediately if removed inside dynamic zone
It's problematic mimicking live's behavior of clearing expedition info on
a timer for cases where a client zones before it triggers. Either the
timer would have to be a non-client based or new zone needs to pick up a
flag in database to do it.

A more complicated solution would be for world itself to handle this but
that would require a deeper redesign