[REBASE] Dire Charm Reset
Added Druid and Necro Dire Charm to Failure Check
In order to make Dire Charm different for the three classes that get it, we had to create identical AAs that linked to similar spells, but for the correct creature type.  Enchanter AA was working (too high of a level would reset AA instead of consume CD), but Druid and Necro were not.