[Spells] Update SPA 339 SE_TriggerOnCast EQEMU PR 1478
Rewrote the functions that handle the focus effect SPA 339 SE_TriggerOnCast
As reminder, the effect works as a focus that gives a chance to cast an additional spell whenever a spell is cast is conditions are met. Base1: Chance Base2: spell_id

This focus was not coded consistently to how other focus effects are checked. There is no longer an arbitrary max number focuses with this effect that a player can have. Overall it works well now, fits with rest of foci code, and is easier to manage.

Also added #command resetdisc_timer
usage: #resetdisc_timer [all | timer_id]