[Spells] Implemented SPA 390 SE_FcTimerLockout EQEMU PR 1496
Implemented

SPA 390 SE_FcTimerLockout

This focus limited effect sets any spell in the spell bar that meets the criteria of the of the focus limits to be a on recast timer.
Base value: recast duration in milliseconds.

Example spell: Suppression of Fire ID 16973.
Sets any fire spell in the clients spell bar to a 2 second recast when the client is affect by the spell.

Note: This focus can only be applied from spells (not item or AA)

Note: Although reinforced by the server, to appear visually correct both server side and client side spell values need to match (ie. need to matching database and spells_us.txt values).