[Bug Fix] Hero Forge armor graphics not displaying properly to other clients in zone.
This update provides fixes to a couple of issues regarding Hero Forge Armor not displaying to other clients in zone under certain conditions.

1- Resolved Issue: When a client wearing the hero forge armor zones into a new zone, other clients in that zone could not see hero forge armor on that client.

2- Resolved Issue: When a client wearing hero forge armor in the helmet removed the helmet, the hero forge graphic would not be reapplied if the item was equipped, until the client zoned. (For some reason we were excluding helmets from wear change, I can see no real reason why, I verified on both titanium AND ROF2 that it has no impact on clients show helm option.)

3- Resolved Issue: When a client wearing hero forge armor removes the item and then re-equipped it in zone, for helmet slot it would not show the hero forge graphic to other clients consistently (even after fixing issue 2), also have reports of other slots having same problem.

For issues 1 and 3, I am not completely certain why the packet is not being received by other clients in zone. I tried many different ways to fix it without luck. But what did work is sending it on slight delay which resolves the problem completely. Might not be the ideal fix, but it is a fix.