Ignore expired lockouts in client getters This makes Client::GetExpeditionLockouts only return unexpired lockouts from the client cache. The cache isn't actively processed so expired lockouts only get cleaned up on a packet update. There's no need to access expired cache entries so the option was removed from all client methods.
This was causing a bug in the group/raid DoesAnyMemberHaveExpeditionLockout
api because it could result in checks against expired lockout entries
(Expedition::HasLockoutByCharacterID and GetExpeditionLockoutsByCharacterID).

This should be the only api that was significantly impacted by this bug.
Other quest apis that went through this method were able to validate
lockout time remaining which is set to 0 when expired.