{EQEmu PR 1989} [Spells] Fix for AA recast timers not resetting properly Reimplementing the portion of the reverted PR that dealt with AA recast timers. Did not appear to be any issues with the AA timers in the reverted PR. #1988
Problem: AA's when they would fail beyond a certain point in the spell casting code would not reset because the timer is set after the spell casting function is completed. The code we have to do the reset doesn't work because the timer doesn't exist yet at the point, so players get the message that the AA was reset, when in fact it was not.

Solution: AA timers are now set within the casting routine after all failure checks are completed.