{EQEMU PR 2030} [AI] Spell Type (1024) InCombatBuff were spam casting
The AI code for spell casting (engaged):

two forms of innate/priority0 spells
in combat buffs
heals/bene
everything else
The first spell to get a sucess roll wins for that trip through the loop.

InCombatBuffs is 50% per spell to be chosen. Hopwever, the code did not have a CanStack check. So mobs with a spell list that included InCombatBuffs is severely compromised due to one spell getting cast over and over, on top if itself, 50% of the loops through if the spell list has no innates/priority0 spells.

Added a simple CanStack check and these mobs behave much better.

(Also changed the brace style to match the surrounding code)