{EQEMU PR 1940} [Content Filtering] Updates contents flags to be checked at runtime.
This helps remove the need to restart server when changing content flags. If you change a flag from enabled to disabled or back, all it requires is a #reloadcontentflags and then a #repop. If you change the flags on a loottable or lootdrop, then you have to also do a #hotfix before #repop.