{EQEmu PR 2037} [Command] Fix #killallnpcs from crashing
This is another command that was crashing from recent cleanups.

Using NPC List is not great for this because it has book keeping issues, I moved it back to using MobList. Using the NPC list was causing crashes in it of itself. It's not a good idea to be iterating over a list that is being manipulated (entries being removed) so in this case using a copy of the mob list works better.