{EQEmu PR 2089} [Compile] Decrease build times using unity build strategy
A while back I introduced 1766 which had the benefit of making our code easier to maintain in terms of splitting commands up, but actually added a significant penalty of compile time - 20-30s to 5m or more depending on hardware.

This is because each cpp becomes its own compilation unit and all of the headers get compiled, parsed for each cpp file and an object get created and it all gets individually linked at the end as well.

This change utilizes a common practice in industry known as Unity Builds

Until C++20 and modules and investing into our project utilizing them, this is a cheap lift to get some development iteration time back.