{EQEmu PR 2095} [Combat] Fix Frenzy vs opponents immune to non-magic
I believe that frenzy should fail against opponents that are immune to non-magical weapons, when not wielding one. This PR fixes that.

NOTE: It seems live has removed this invulnerability from wisps and ghouls... Not sure if it's just gone, or only on selected mobs.

Also note, this fixes the issue for clients. Very similar code exists in Client::DoClassAttacks for Frenzy. Does this need to be changed also? I don't know where that code is called from?? Changing that code had no effect on this bug. Maybe I need to put the same fix there for cases when client is AI controlled?