{EQEmu PR 2186 } [Login] Added OP_ExpansionPacketData for RoF2 and update payload for Titanium
Good afternoon.

I have added the ability to update the expansion info on the Server Select screen for Titanium and Rof2. Titanium uses the encrypted payload while RoF2 uses a new packet structure.

I have based it on using the rule World:ExpansionSettings and force the client to stay within it's max as it is based on data within eqlsstr_us.txt.