{EQEmu PR 2246} [Fix] Boats should never get FixZ'd
If a boat spawn point or waypoint had a Z value that caused the code to no longer see the boat in water, even though a great deal of the boat was below water. FixZ would kick in. This can cause massive issues to clients viewpoints of the boat (other than the client riding the boat) - causing some people that zone in after the FixZ to see the boat and rider underwater.

This is a tiny patch to avoid the issue in general - but I'll be looking to touch up my Boat navigation in movement_manager to be more in sync with how the client deals with Z for boats.