{EQEmu PR 2277} [Bug Fix] NPCs were getting weapon proc added twice
NPCs and pets that wield a weapon with a proc were getting the proc added twice to the in memory tables.

This also caused pets to sometimes proc weapons that should not proc for their level, while sometimes yielding the "not able to command this weapon" other times.

Tested thoroughly with the fix.