{2289} [DoT Messages] Add DoT messages for mob->PC casts, fixed others to use correct strings
DoT's cast by mobs on PCs did not produce the messages. Fixed this, and fixed duel situations as well.

Needed to extend FilteredMessageCloseString to be able to ignore one client in addition to the sender.

Fixed DoT to use the injured party as center for the CloseClients part instead of the caster as far as broadcasting.