{EQEmu PR 2302} [Logs] #logs list Improvements
This PR adds improvements to the `#logs list` command that adds clickable saylinks for each log category.

If the logging level is already active, it is shown without a saylink as a white number, otherwise it is printed as a clickable saylink to activate the logging category for that category.

This does remove the popup as your focus will remain in the chat window for listings options. This still keeps paging as it was

**In-Game Example**

![image](https://user-images.githubusercontent.com/3319450/178127555-3c5ec77c-b7d8-442c-835e-2da98468c8af.png)