{EQEmu PR 2727} [Diawind] Plus sign markdown fix
### Original Problem

"Balloon
 — Today at 9:35 AM
I have a DiaWind box that explains some bonuses (e.g. +10 STR), but the + is being interpreted as an animation command and the display gets very screwy.  Is there a way to show a specific character without it being interpreted?  I've tried prefixing with the usual special chars and also using html &#43; to no avail."

### Fix

We only replace content between plus signs when an animation is found