{EQEmu PR 2708} [Rule] Add rule for NPC Level Based Buff Restrictions.
Added a rule that allows server ops to apply "SPELL_TOO_POWERFUL" rule to NPCs just like PCs.

NPCBuffLevelRestrictions (default off to retain existing functionality)

While debuging this I found that NPC spells from the AI that fail in the CastSpell() portion of the spell casting chain do not recover ever - as the code relies on the casting_spell_id  and the spell slot to be valid in every function that calls the AI with AI_Event_SpellCastFinished.  As these variables are not set until later in the casting chain, the  AI was never getting reset.

This bug was much more obvious when setting my new rule, as NPC spells fail in the checking stage alot more.

This is a DRAFT PR until someone more adept with spells than I ok's the change.

I did test all the variant cases and it works, and I put a log locally if an NPC ever cast a spell not using slot 1 and I didnt see any.  Not sure that matters, but the AI only resets on slot 1, and since slot was not set, I pass in 1.