{EQEmu PR 2766} [Logging] Remove function prefixes
What
With PR's #2765 and #2755 brings code function prefixes to logging statements automatically. This removes the need to manually have them in our logging statements all over the code and in many cases will have double of the same information with the prior logging changes.

This PR removes most places in the code where the function is inserted as part of the logging statement