{EQEmu PR 2821} [Doors] Have NPCs trigger double doors
### What

This PR solves for a situation where NPC's clip through double doors because they have a trigger door and the current AI does not let them open doors that have a trigger door. This is usually because we don't want NPC's triggering for example a lever that elevates a platform.

We do want NPC's to be able to open double doors

### Solution

When an NPC walks into a door that has a trigger door, compare the names without numbers to see if they are complimentary. In majority cases double doors are similarly named but have a different number suffix.

This automatically solves for the following zones that fall under a similar pattern (https://gist.github.com/Akkadius/cfbc826c193b33a1c4687fd0001fb915)

* gfaydark
* soltemple
* citymist
* hole
* paineel
* sleeper
* katta
* potactics
* potorment
* tutorialb
* ikkinz

Example doors

```
select id, doorid, zone, `name`, triggerdoor from doors where doors.zone = 'soltemple' and doorid in (43, 44);
+------+--------+-----------+----------------+-------------+
| id   | doorid | zone      | name           | triggerdoor |
+------+--------+-----------+----------------+-------------+
| 6249 |     43 | soltemple | SOLTEMPDOOR400 |          44 |
| 6248 |     44 | soltemple | SOLTEMPDOOR401 |          43 |
+------+--------+-----------+----------------+-------------+
2 rows in set (0.00 sec)
```

### Demo

https://user-images.githubusercontent.com/3319450/216173181-365f1fd8-2424-4b19-94ac-1b7fb6b81e9f.mp4