{EQEmu PR 2827} [Tasks] Implement alternate currency rewards
### What

Implements alternate currency rewards. Spire already had implementation for this but we only handled Ebon / Radiant in source. This may not be specifically how live displays the reward but it gets the functionality in place until it is adjusted later. Right now we're only up until Dragons of Norrath content wise which does not care about alternate currencies so this is more focused to custom servers.

![image](https://user-images.githubusercontent.com/3319450/216803873-bb731a33-8599-4020-97a2-ca75425d7795.png)

### Solution

This currently utilizes the `reward_point_type` and `reward_points` fields to produce an item link of the alternate currency. It takes the place of what would otherwise be a simple item reward display in the task window.

![image](https://user-images.githubusercontent.com/3319450/216803899-dffa422c-3f4f-4768-8e52-b1c22f604d22.png)

### On Reward (Task Completion)

![image](https://user-images.githubusercontent.com/3319450/216803934-1e5e388e-fb8b-4349-b6d3-51617ce6b6ce.png)

![image](https://user-images.githubusercontent.com/3319450/216804327-3f0ce0e1-a55f-42b6-9e13-f4aca3f38359.png)