{EQEmu PR 2867} [Crash] Fix Crash in Mob::FindType
Limited checks were being performed in Mob::FindType to determine if returned spell_id was valid. obviously we should also try to determine why an invalid spell ID can be entered into Buffs_Struct in the first place. But I wanted a fix to resolve the crashing seeing on karana.