{EQemu PR 2932} [Tradeskills] Check if Combine would result in Lore Conflict.
We previously were not checking if the results of a combine would cause a Lore Conflict, and resulting items would be eaten.

Message Client receives now, when there is a Lore Conflict, and preventing the combine from occurring.

Added Item Links
![image](https://user-images.githubusercontent.com/109764533/219112596-2e007cf1-ae68-4a88-8b16-01ae8c53eae6.png)

(I have no idea what is with the formatting on string_ids.h everything I do seems to make it worse, but it looks fine in my IDE.