{EQEmu PR 2953} [MySQL] Add keepalives to UCS and Loginserver
### What

This adds MySQL keepalives to both UCS and Loginserver to MySQL. Zone and World already have them and have had them for a long time.

This hasn't posed any issues but recently @noudess reported an issue with his MySQL 8 instance and it appears MySQL 8 is going to be a bit more sensitive.

This PR keeps the connection fresh by issuing pings every 10 seconds

### Reported Issue

Noudess "This message always comes out on 1st login attempt after an account has been logged out for long periods of time.  Can someone help me find where this is in the code?  I searched for some of this text and cannot find it.  From just the error message, it feels like some previous logout was missed and a new attempt to login triggers the error.  After 1 failure to login, all works again as normal.

Message:

[02-16-2023 10:34:38] [Login] [QueryErr] [4031] [The client was disconnected by the server because of inactivity. See wait_timeout and interactive_timeout for configuring this behavior.]
[SELECT id, account_password FROM login_accounts WHERE account_name = 'Paul3' AND source_loginserver = 'local' LIMIT 1]"