{EQEmu PR 2975} [Pathing] Improvements to z-clipping, z-recovery and z-calculations
### What

Augments https://github.com/EQEmu/Server/pull/2826 by adding back in the original `z_find_offset` (default 5) to the Z find. This has made Z finding more reliable on PEQ in conjunction to the previous changes.

This also moves Z clip detection out of the Z calculation function and moves it to a 1-second timer that happens while the NPC is engaged.

If the NPC is close to the player in X/Y but the Z has fallen out of LOS, the NPC will recover to the player. This has worked remarkably well.

In conjunction with this, navmesh changes have been made to various zones to improve the mesh as we still have very gappy zones. See http://www.projecteq.net/forums/index.php?threads/server-update-2-20-2023.17509/ for example