{EQEmu PR 2985} [Doors] Fix doors triggering invalid zone fetches of dest_zone of "none"
### What

@noudess reported his zone was spamming

```
 Zone |    Info    | operator() New client from [192.168.0.5:62944] -- [paineel] (Paineel) inst_id [0]
 Zone |    Info    | GetZoneID Failed to get zone_name [none] -- [paineel] (Paineel) inst_id [0]
```

This sources from door logic when opening a door and the code is not properly checking for "none". The code makes attempts to do so but my guess is some sort of case sensitivity / collation is occurring with the database.

This PR attempts to clean some of that up and make the code a bit more readable.

This PR also needs to be tested to ensure no regressions