{EQEmu PR 2987} [Bug Fix] Fix Beneficial Target of Target procs
Beneficial Procs that were of Target Type ST_TargetsTarget were attempting to be cast on the Players Target, and not the Player's Targets Target, as we were passing the spells target as this for all Beneficial proc spells.

This changes the behavior to correctly pass the spell target as the players current target, which will result in the correct Target being used.