{EQEmu PR 2982} [Pathing] Smoother pathing z-correction
### What

I was testing some unrelated stuff and noticed that when I drug NPC's up and down hills that they would "hop" only when I would change position. This tells me that the server is sending a new position update when I change position and with it comes the current Z that the server has for the NPC.

You don't notice the hopping issue if the NPC is already running towards you and you didn't move because this is purely a client animation and the client has already done its corrections client side to properly plant the NPC on the ground.

After testing, I found that we were not calculating the "Fixed Z" fast enough during pathing, especially while engaged. It was previously set at 10 units per a discussion KLS and I had years ago to keep the expensive calculations down but this appears to cause visual problems which cheapens the appearance of our pathing mechanics.

With this change - when an NPC is engaged, the calculation interval is high resolution (1 unit) otherwise when not engaged the Z calculation is done at the original 10 units

**Before**

https://user-images.githubusercontent.com/3319450/220830880-7278772b-33d4-412a-997e-56d04093ac4a.mp4

**After**

https://user-images.githubusercontent.com/3319450/220831271-9605d4dd-8535-43d4-b82a-44b98d950a52.mp4