{EQEmu PR 2834} [Bug Fix] Fix for undefined MySQL library behavior.
The MYSQL structure in the MySQL C api is not copy aware. This isn't enforced because, well, C doesn't enforce anything really.
See: MySQL Structures

This changes the copy of an existing connection to instead create a second backup connection with the same data as what is being copied. This does introduce another connection into the mix which are a limited resource but given the potential seriousness of the bug in question it seems like a fair tradeoff for now.

Long term we can rewrite the dbcore class to use a underlying pointer instead of a pinned in place variable.