{EQEmu PR 2988} [Pathing] More z-clip improvements, Wurm and Spectral Iksar race adjustments
### What

* Improve z-clipping logic, use Z offset calculations
* Z-clipping reset no longer resets NPC "guard" location (spawn point)
* Wurms (Race 158) as a fixed size model are now statically set to size 15 on spawn to make up for data issues
* Adjust Spectral Iksar (Race 147) offset causing an edge case where it would clip into the ground just below the surface mesh in Trakanon

### Not returning to spawn point

Below the NPC does not return to spawn point after z-clip correction, this has been fixed.

https://user-images.githubusercontent.com/3319450/221330680-0506a994-a5ad-4025-9b66-d66ad6588a8d.mp4

#### Wurm Clipping

Wurm is falsely detected as under or below player simply because the size of the NPC is not taken into consideration in relation to the player. Both player and NPC offsets are taken into account and this no longer occurs.

![3be7b16806b5a4b8c68973a313c235c4-min](https://user-images.githubusercontent.com/3319450/221330554-131cc61d-7d9c-4a9d-84de-f10a0aa78f61.gif)

#### Spectral Iksar Issue

NPC's model offset was adjusted and after 10-20 minutes I've not observed any clips into the geometry below the top floor

![cda9f7eb1ba6e0c42769eca8d6fe20f7-min](https://user-images.githubusercontent.com/3319450/221330462-acadf786-5db9-4938-8dfb-2145a297d62b.gif)