{EQEmu PR 3005} [Bug Fix] Fix for Lore Components where component is returned.
This resolves issues seen with BiC quest line combines, where the item is returned on success, fail, salvage. However the recipe entries for each different return type are split between multiple rows.

My previous PR was meant to fix the above, but was incorrect logic. my testing failed to find this, as the checks worked correctly on the first component if it was lore, not later components.

The correct implementation is to loop through all components first before performing any lore checks, allowing us to exclude any components that are lore, but have a separate row for return on success, fail, salvage.

Verified all previous use cases work as expected as well, meaning:

- Combine uses a lore container, which is returned.
- Combines that have a Lore Group component (Epic Lore Group, or other above 0) and return the same lore-group on success.
- Components are lore, and returned on success, fail, salvage (but in same row entry)