{EQEmu 3008} [Bug Fix] Fix for Discipline Loading from Database causing issues with slot_ids
If a Server Operator attempts to cleanup disciplines after they are learned from the database, they may encounter issues after the fact when players try to learn new disciplines due to how we were loading data into the Disciplines_Struct.

We were loading data in with no accounting for Slot_id, so when learning a new Discipline if an older row was deleted the Struct would not match up with the correct Slot_id numbers, causing players to overwrite disciplines they have already learned.

This allows a null entry to be added to the Disciplines_Struct.disciplines.value, so when we learn a new discipline it'll find the null and insert to the slot that is free, and not try to insert at the end.