{EQEmu PR 3021} [Console] Add IS_TTY to force terminal coloring output
### What

Add `IS_TTY` to force terminal coloring output

This is used to enable outside processes to force terminal output to be colored. For example, Spire Admin panel has a pre-flight checks that is ran to ensure the server is in working order and has no errors. The output gets sent as ANSI and parsed from ANSI to HTML to give the colored resemblance of what you would expect to see in the console.

![image](https://user-images.githubusercontent.com/3319450/222841898-68d555d3-a71e-4d8f-8814-243b8023782e.png)