{EQEmu PR 3011} [Bots] Prevent interrupt spam when OOM
Notes:

https://user-images.githubusercontent.com/53322305/223123434-2bcba221-d13f-4320-9995-e6a3619ccb84.mp4

Bots currently do not do a mana check until the spell cast has already been started which results in an interrupt immediately after. If it is the last spell for a bot to cast, it tends to result in interrupt spam until the bot has enough mana. This will block the interrupt spam if the spell is invalid.