{EQEmu PR 3151} [Crash] Add Checks for out of bounds & dereferencing nullptrs
Added checks to prevent crashes as seen in the below instances:

http://spire.akkadius.com/dev/release/22.8.0?id=2826
http://spire.akkadius.com/dev/release/22.8.0?id=2830

Mob::SpellEffect needs refactored eventually, but these checks should help ensure we don't call an out of bounds memory address.

buffslot can be either -2, -1 or a positive value (that should be valid) so anywhere we use buffslot to access buffs[buffslot] we need to ensure there's a check to ensure we're not out of bounds.