{EQEmu PR 3175} [Backups] Fix database dump error reporting
### What

This PR fixes an issue where database dumps would not bubble `stderr` from the underlying process execution. Currently this pipes `stderr` to `stdout` so we don't miss errors, this could be later improved to hold a response object containing `status` `stderr` and `stdout`

This is to prevent server operators from experiencing any sort of data loss if they are missing a table that the database backup is trying to request explicitly. If a table is missing, `mysqldump` will normally stop.

We could ignore this but this really is something that should be addressed immediately otherwise it could cause worse pain to ignore.

![image](https://user-images.githubusercontent.com/3319450/229331154-c8ab0adc-3da1-4cb8-a6cd-25790f91b081.png)

* We now cleanup SQL files after compression is requested and a compression binary is found
* When a `mysqldump` error is detected, the world process is exited forcefully
* Added file utility `File::GetContents(file_name)`

![image](https://user-images.githubusercontent.com/3319450/229331923-b484478a-390f-4381-85d9-4811fee7fef4.png)