{EQEmu PR 3296} [Crash] Fix crash with uninitialized item instance, and Bot timeout
This PR fixes a crash where an uninitialized item instance is called (this forces a nullptr on initialization), this also fixes a "crash" where the zone would get stuck in a while loop.

I believe both of these are regression issues, however the while loop has been in place for a very significant time (years) it's likely that this code was never actually being hit before.

http://spire.akkadius.com/dev/release/22.9.1-dev?id=3186