{EQEmu PR 3136 & 3318} [Feature] Add support for -1 extradmgskill to allow all skills to be scaled. [BugFix/NPCs] Fix issue with NPCs no longer using some armor.
# Perl
- `$mob->GetSkillDmgAmt(skill_id)` now uses `int` instead of `uint16`.

# Lua
- `statbonuses:GetSkillDamageAmount(skill_id)` now uses `-1` properly.
- `mob:GetSkillDmgAmt(skill_id)` now uses `int` instead of `uint16`.

# Notes
- A `-1` value in `extradmgskill` denotes the ability to scale all secondary combat skills at once.
- Consolidated `AddItemBonuses()`, `AdditiveWornBonuses()`, `CalcItemBonuses()`, and `CalcRecommendedLevelBonus()` to mob-based methods to avoid code duplication.
- Bots, NPCs, and Mercs can now use additive worn effects if the rule is enabled, as well as all other proper stat bonuses that only clients had before.
- No SQL update required to change `extradmgskill` and `extradmgamt` to `int` as they already are this type in the database, just had to adjust `item_data.h` and `shareddb.cpp`.

# Skills Affected
```cpp
std::vector<int> secondary_combat_skills = {
	EQ::skills::SkillBackstab,
	EQ::skills::SkillBash,
	EQ::skills::SkillDragonPunch, // Same ID as Tail Rake
	EQ::skills::SkillEagleStrike,
	EQ::skills::SkillFlyingKick,
	EQ::skills::SkillKick,
	EQ::skills::SkillRoundKick,
	EQ::skills::SkillTigerClaw,
	EQ::skills::SkillFrenzy
};
```

# Images
![image](https://user-images.githubusercontent.com/89047260/226792508-47407d85-1bfa-44b7-9ca2-b56d23464786.png)
![image](https://user-images.githubusercontent.com/89047260/226792521-c09f8adb-07e0-4887-a5d5-afb2068dccb3.png)

This PR is to resolve a problem brought up [here](https://discord.com/channels/212663220849213441/1098664193357520996).

The issue is that pets were often not using armor traded to them any longer because the Bonuses code had been consolidated into a single module and that module was applying some client restrictions like race/class, which cannot pass for pets. (As an aside, the same problem was created for NPCs of various races).

That consolidation was done in [PR](https://github.com/EQEmu/Server/pull/3136).

While working on this, I also found a longer term problem in that an item traded to a pet was not actually impacting that pet's AC or other stats until another event which would impact bonuses occurred. The older code did not have this issue, as NPCs were using their internal equipment array which was updated when AddLootDrop() called CalcBonuses.  Now, however, the combined Bonuses code uses the itemlist via the NPC.h interface.  So the call to CalcBonuses needed to be moved below where itemlist was updated in AddLootDrop().

I also cleaned up and if/else section in bonuses.cpp where one if is cleaner.