{EQEmu PR 3323} [Messages] Remove duplicate you have lost a level message
I have verified on titanium and rof2 that we always get 2 "you have LOST a level" messages and that the one in yellow is from OUR code, the one in RED I can find nowhere so I am thinking the client is generating it.

I cannot find anywhere in our code where the constant LOSE_LEVEL is used other than SetExp and nowhere do we send the entire message hard coded or use the value 442 in a message.

Therefore I think we can remove this message and rely on the client.